refactor(l5): agent-context-pack, durable-lane and durable-memory become L5 adapters over their server/tools twins (issue 864)
Three src/tools context-pack modules already had server/tools twins, so per M16
they become adapters over those twins rather than being moved. The twins were
found by exported symbol across server/, not by filename.

Adapter mapping (src path -> server twin it now serves):

  src/tools/agent-context-pack.ts
    -> server/tools/agent-context-pack.ts
       M9 adapter: keeps the legacy ToolDeps call form and maps it to the
       twin's MemoryToolDependencies shape inside the adapter.

  src/tools/agent-context-pack-durable-lane.ts
    -> server/tools/context-pack-durable-lane.ts
       M9 adapter, same deps mapping.

  src/tools/agent-context-pack-durable-memory.ts
    -> server/tools/context-pack-durable-memory.ts
       M2 shim: the legacy positional wrapper had no caller in the repo, so
       the file is a plain re-export rather than an adapter.

Legacy dependency relocation: the environment-derived values the twins take as
configuration now live in server/tools/legacy-context-pack/legacy-deps.ts, which
the adapters fill. server/tools/legacy-context-pack/agent-context-pack-pointers-candidates.ts,
server/application/nats-runtime.ts, server/tools/realtime-append.ts and
src/realtime/recovery-wal.ts follow that relocation.

M5 test transformations applied to the moved suites:
  - every non-null assertion became expectDefined(value, label) from
    scripts/test-support/expect-defined.ts
  - the durable_memory failure-path structural guard was repointed to the
    twin's source path and to logDurableFailure, the shared two-line failure
    shape the twin emits

scripts/done-means/744-recall-serves-durable.sh was rewritten in this commit per
M16: its two probes no longer compare a src implementation against a server one,
because at L5 both resolve to the same twin through the two entrypoints callers
actually reach. The comment says the comparison collapsed at L5. The driver still
reports RESULT: FAIL, which is the pre-existing issue 744 behaviour it was
written to detect (a bare recall does not consult durable memory) and not a
regression from this change.

M14 test split: src/tools/__tests__/agent-context-pack-durable-memory.test.ts
failed the whole-file pre-commit lint at 1035 code lines with a 987-line describe
callback. Per M14 it is split mechanically in this same commit, by its own it()
groups, with assertions and it-names unchanged. The original file is renamed with
git mv to the first group so history follows.

  file                                                     lines  tests
  agent-context-pack-durable-memory-empty-states.test.ts     160      5
  agent-context-pack-durable-memory-isolation.test.ts        158      3
  agent-context-pack-durable-memory-allocation.test.ts       348      5
  agent-context-pack-durable-memory-recall-failure.test.ts   171      3
  agent-context-pack-durable-memory-suppression.test.ts      340      7
  agent-context-pack-durable-memory-determinism.test.ts      143      3
  agent-context-pack-durable-memory-test-helpers.ts           70      -
                                                                    ---
                                                                     26

The shared searchPool, brainRecord and JsonRecord fixtures moved into
agent-context-pack-durable-memory-test-helpers.ts, following the sibling
precedent agent-context-pack-durable-lane-test-helpers.ts. The same 26 tests run,
with the same names and the same assertions. scripts/assert-db-tests-ran.ts names
none of these files (they are not *.pg.test.ts), so no manifest edit was required.

Closure of src/ under server/main.ts: 2 files (shims and adapters are excluded by
the header regex per M9).

Verified:
  bunx tsc --noEmit                                              -> exit 0
  oxlint --deny-warnings on every staged path                    -> exit 0
  bun run test:isolated agent-context-pack-durable-memory-*      -> 26 pass 0 fail
  bun run test:isolated agent-context-pack*                      -> 146 pass 0 fail, 18 files
  bun contracts/check-parity.ts                                  -> exit 0
  bash scripts/done-means/864-moved-out-of-src.sh                -> PASS, judged=48 shims 17 adapters
